refactor virtual list implementation
