+++ src/hooks/useFlattenRows.ts
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import type { Group, GroupSegmentItem } from './useGroupSegments';
+import type { Group } from '../interface';
+import type { GroupSegmentItem } from './useGroupSegments';
 
 export type Row<T, K extends React.Key = React.Key> =
   | { type: 'header'; groupKey: K }
