+++ src/hooks/useGroupSegments.ts
@@ -1,9 +1,7 @@
 import * as React from 'react';
+import type { Group } from '../interface';
 
-export interface Group<T, K extends React.Key = React.Key> {
-  key: (item: T) => K;
-  title: (groupKey: K, items: T[]) => React.ReactNode;
-}
+export type { Group } from '../interface';
 
 export interface GroupSegmentItem<T> {
   item: T;
