+++ tests/hooks.test.tsx
@@ -4,8 +4,8 @@ import type { ListProps as VirtualListProps } from '@rc-component/virtual-list';
 
 import useFlattenRows from '../src/hooks/useFlattenRows';
 import useGroupSegments from '../src/hooks/useGroupSegments';
-import useStickyGroupHeader from '../src/hooks/useStickyGroupHeader';
-import type { StickyHeaderParams } from '../src/hooks/useStickyGroupHeader';
+import useStickyGroupHeader from '../src/VirtualList/useStickyGroupHeader';
+import type { StickyHeaderParams } from '../src/VirtualList/useStickyGroupHeader';
 
 const PREFIX_CLS = 'rc-listy';
